Fix Banco de Dados

+++ LocadoraVeiculo/LocadoraVeiculo.Database.SQLServer/dbo/Tables/TBTipoCombustivel.sql
@@ -1,6 +1,8 @@
 ﻿CREATE TABLE [dbo].[TBTipoCombustivel] (
-    [Preco_Gasolina] FLOAT (53) NULL,
-    [Preco_Diesel]   FLOAT (53) NULL,
-    [Preco_Alcool]   FLOAT (53) NULL
+    [Preco_Gasolina] DECIMAL (18) NULL,
+    [Preco_Diesel]   DECIMAL (18) NULL,
+    [Preco_Alcool]   DECIMAL (18) NULL
 );
+
+
